Continuing developing CMS, adding, edit and fixing features
Adding a message when using succesfully adding an information. (user & admin post page)

Update a post status's select tag for edit & add page

Adding a new bar graph in dashboard for viewing an active post (Published post).

Using a CKEditor plugin for a much powerful textarea area tag for admin post pages.

Adding a link to the image and read more button in the main index page, so that it will bring user to that  post's page.

Adding Edit post button for admin to edit that spesific post in the post's page not the admin on, this one using session to detect user roles.

Fixing a problem with session, by adding session_start function in the header of the front page

Adding a template for bulk option in view post, there's no functionality yet.