Implemented real-time notification using pusher, toastr
using pusher to notify admin that  new users just registered to the application

Combining pusher with toastr to outputting message dynamically and beautifully possible.

Setting .env in the project folder to protect the credential, key and etc.

Notes; this method, hopefully can be added to other functionality of the system in the future, like adding a new user, editing user and etc.